search-manager-agent/requirements.txt
autogen-agentchat>=0.4.0
autogen-ext[openai]>=0.4.0
python-dotenv>=1.0.0
aiohttp>=3.9.0
